% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distance_method.R
\docType{methods}
\name{distance}
\alias{distance}
\alias{distance,matrix-method}
\alias{distance,simple_triplet_matrix-method}
\title{Calculate Distance}
\usage{
distance(x, ...)

\S4method{distance}{matrix}(x, method = "jaccard", verbose = TRUE)

\S4method{distance}{simple_triplet_matrix}(x)
}
\arguments{
\item{x}{input matrix}

\item{method}{defaults to "jaccard", in fact, this is the only method that is implemented}
}
\value{
object of class \code{dist}
}
\description{
Accelerated way using matrix multiplication for jaccard distance computation.
http://stats.stackexchange.com/questions/49453/calculating-jaccard-or-other-association-coefficient-for-binary-data-using-matri
}

