% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_method.R
\name{divide}
\alias{divide}
\alias{divide,matrix-method}
\alias{divide,vector-method}
\title{Divide an object into roughly equally sized parts.}
\usage{
divide(.Object, n)

\S4method{divide}{matrix}(.Object, n)

\S4method{divide}{vector}(.Object, n)
}
\arguments{
\item{.Object}{object to split up}

\item{n}{number of objects}
}
\value{
a list of objects of the same class as the input object
}
\description{
This is a helper to prepare lapply/mclapply operations on vectors, or matrices.
}
\examples{
divide(1:10, n = 3)
M <- matrix(1:20, ncol = 2, byrow = TRUE)
divide(M, n = 3)
divide(M, n = 4)
}
\seealso{
\code{\link[text2vec]{split_into}} in the text2vec
package offers a similar, and faster function for vectors
}
