% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicates_class.R
\name{Duplicates}
\alias{Duplicates}
\title{Detect Duplicates}
\description{
Detect Duplicates

Detect Duplicates
}
\details{
Class for duplicate detection.

The class implements a procedure described by Fritz Kliche, Andre Blessing,
Urlich Heid and Jonathan Sonntag in the paper "The eIdentity Text
ExplorationWorkbench" presented at LREC 2014
(see \url{http://www.lrec-conf.org/proceedings/lrec2014/pdf/332_Paper.pdf}).

To detect duplicates, choices are made as follows:
\itemize{
\item If two similar articles have been published on the same day, the shorter article will
be considered the duplicate;
\item if two similar articles were published on different days,
the article that appeared later will be considered the duplicate.
}

Different \code{partition_bundle}-objects can be passed into the
\code{detectDuplicates}-method successively. The field \code{duplicates} will be
appended by the duplicates that are newly detected.
}
\examples{
library(polmineR)

if ("NADIRASZ" \%in\% corpus()$corpus){
  coi <- "NADIRASZ"
  s_attr_date <- "article_date"

  D <- Duplicates$new(
    char_regex = "[a-zA-ZäöüÄÖÜ]",
    p_attribute = "word",
    s_attribute = s_attr_date,
    datePrep = NULL,
    sample = 50L,
    n = 1L,
    threshold = 0.6 # default is 0.9
  )

  article_bundle <- corpus(coi) |>
    subset(article_date == "2000-01-01") |> 
    split(s_attribute = "article_id")

  D$detectDuplicates(x = article_bundle, mc = 3L)
  
  # To inspect result
  D$duplicates
  
  if (interactive()){
    for (i in 1L:nrow(D$duplicates)){
    
      print(i)
      
      corpus("NADIRASZ") \%>\%
        subset(article_id == !!D$duplicates[i][["name"]]) \%>\%
        read() \%>\%
        show()
        
      readline()
  
      corpus("NADIRASZ") \%>\%
        subset(article_id == !!D$duplicates[i][["duplicate_name"]]) \%>\%
        read() \%>\%
        show()
        
      readline()
    }
  }
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{corpus}}{ID of the CWB corpus (derived from \code{partition_bundle}).}

\item{\code{char_regex}}{Regular expression defining the characters to keep.}

\item{\code{char_count}}{Count of the characters in the \code{partition_bundle}.}

\item{\code{n}}{Number of days before and after a document was published.}

\item{\code{p_attribute}}{the p-attribute used (defaults to "word")}

\item{\code{s_attribute}}{the s-attribute of the date of a text in the corpus}

\item{\code{sample}}{size of the sample of the \code{partition_bundle} that the character count is based on}

\item{\code{threshold}}{minimum similarity value to identify two texts as duplicates}

\item{\code{comparisons}}{a \code{simple_triplet_matrix} with the texts to be compared}

\item{\code{similarities}}{a \code{simple_triplet_matrix} with similarities of texts}

\item{\code{ngramDocumentMatrix}}{a matrix (inheriting from \code{TermDocumentMatrix}) with ngram counts in the documents of the \code{partition_bundle}}

\item{\code{datePrep}}{function to rework dates if not in the DD-MM-YYYY standard format}

\item{\code{annotation}}{a \code{data.table} with corpus positions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Duplicates-new}{\code{Duplicates$new()}}
\item \href{#method-Duplicates-getWhatToCompare}{\code{Duplicates$getWhatToCompare()}}
\item \href{#method-Duplicates-makeDuplicateDataTable}{\code{Duplicates$makeDuplicateDataTable()}}
\item \href{#method-Duplicates-detectDuplicates}{\code{Duplicates$detectDuplicates()}}
\item \href{#method-Duplicates-makeAnnotation}{\code{Duplicates$makeAnnotation()}}
\item \href{#method-Duplicates-encode}{\code{Duplicates$encode()}}
\item \href{#method-Duplicates-clone}{\code{Duplicates$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Duplicates-new"></a>}}
\if{latex}{\out{\hypertarget{method-Duplicates-new}{}}}
\subsection{Method \code{new()}}{
Initialize object of class \code{Duplicates}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Duplicates$new(
  char_regex = "[a-zA-Z]",
  p_attribute = "word",
  s_attribute = "text_date",
  datePrep = NULL,
  sample = 1000L,
  n = 1L,
  threshold = 0.9
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{char_regex}}{a regex defining the characters to keep}

\item{\code{s_attribute}}{the s-attribute providing the date}

\item{\code{sample}}{number of documents to define a subset of \code{partition_bundle} to
speed up character count}

\item{\code{n}}{number of days before and after a document was published}

\item{\code{threshold}}{numeric (0 < x < 1), the minimum similarity to qualify two documents as duplicates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Duplicates-getWhatToCompare"></a>}}
\if{latex}{\out{\hypertarget{method-Duplicates-getWhatToCompare}{}}}
\subsection{Method \code{getWhatToCompare()}}{
Identify documents that will be compared (based on date of documents).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Duplicates$getWhatToCompare(
  x,
  reduce = TRUE,
  verbose = FALSE,
  progress = TRUE,
  mc = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a \code{partition_bundle} object defining the documents that will be
compared to detect duplicates}

\item{\code{verbose}}{logical, whether to be verbose}

\item{\code{progress}}{logical, whether to show progress bar}

\item{\code{mc}}{logical, whether to use multicore}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Duplicates-makeDuplicateDataTable"></a>}}
\if{latex}{\out{\hypertarget{method-Duplicates-makeDuplicateDataTable}{}}}
\subsection{Method \code{makeDuplicateDataTable()}}{
Turn similarities of documents into a data.table that identifies original
document and duplicate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Duplicates$makeDuplicateDataTable(
  x,
  mc = FALSE,
  progress = TRUE,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a \code{partition_bundle} object defining the documents that will be
compared to detect duplicates}

\item{\code{mc}}{logical, whether to use multicore}

\item{\code{progress}}{logical, whether to show progress bar}

\item{\code{verbose}}{logical, whether to be verbose}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Duplicates-detectDuplicates"></a>}}
\if{latex}{\out{\hypertarget{method-Duplicates-detectDuplicates}{}}}
\subsection{Method \code{detectDuplicates()}}{
Wrapper that implements the entire workflow for duplicate detection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Duplicates$detectDuplicates(
  x,
  n = 5L,
  character_selection = 1:12,
  how = "coop",
  verbose = TRUE,
  mc = FALSE,
  progress = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A \code{partition_bundle} or \code{subcorpus_bundle} object.}

\item{\code{n}}{The number of characters to use for shingling (\code{integer} value),
passed as argument \code{n} into \code{polmineR::ngrams()}. Defaults to 5, in
line with Kliche et al. 2014: 695.}

\item{\code{character_selection}}{Numeric/integer vector used for indexing
\verb{$char_count} to select the characters to keep. Defaults to 1:12, in
line with Kliche et al. 2014: 695.}

\item{\code{how}}{Implementation used to compute similarities - passed into
\code{cosine_similarity()}.}

\item{\code{verbose}}{logical, whether to be verbose}

\item{\code{mc}}{logical, whether to use multicore}

\item{\code{progress}}{logical, whether to show progress bar}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Duplicates-makeAnnotation"></a>}}
\if{latex}{\out{\hypertarget{method-Duplicates-makeAnnotation}{}}}
\subsection{Method \code{makeAnnotation()}}{
Turn data.table with duplicates into file with corpus positions and
annotation of duplicates, generate cwb-s-encode command and execute it,
if wanted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Duplicates$makeAnnotation(sAttributeID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Duplicates-encode"></a>}}
\if{latex}{\out{\hypertarget{method-Duplicates-encode}{}}}
\subsection{Method \code{encode()}}{
´
Add structural attributes to CWB corpus based on the annotation data that
has been generated (data.table in field annotation).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Duplicates$encode(
  exec = FALSE,
  filenames = list(duplicate = tempfile(), original = tempfile())
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Duplicates-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Duplicates-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Duplicates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
