% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hitanalysis_class.R
\docType{data}
\name{Hitanalysis}
\alias{Hitanalysis}
\title{Analyse hits}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Hitanalysis
}
\description{
A story to be told.
}
\keyword{datasets}

