% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarity_method.R
\name{jaccard_distance}
\alias{jaccard_distance}
\title{Calculate Jaccard Distance}
\usage{
jaccard_distance(X)
}
\arguments{
\item{X}{input matrix}
}
\description{
Fast way by using matrix multiplication.
}
\details{
http://stats.stackexchange.com/questions/49453/calculating-jaccard-or-other-association-coefficient-for-binary-data-using-matri
}

