% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label_method.R
\docType{methods}
\name{label}
\alias{label}
\alias{label,partitionBundle-method}
\title{assign labels}
\usage{
label(.Object, ...)

\S4method{label}{partitionBundle}(.Object, labels = c(true = 1, false = 0),
  description = "Make your choice", logfile = NULL, resume = FALSE, ...)
}
\arguments{
\item{.Object}{an object of class \code{"partitionBundle"}}

\item{labels}{labels to assign}

\item{description}{output}

\item{logfile}{a file}

\item{resume}{logical}

\item{logfile}{a logfile}
}
\description{
assign labels
}
\examples{
\dontrun{
use("polmineR.sampleCorpus")
all <- partition("PLPRBTTXT", list(text_id=".*"), regex=TRUE, type="plpr")
speeches <- as.speeches(all, sAttributeDates="text_date", sAttributeNames="text_name", gap=500)
speechesSample <- sample(speeches, 25)
df <- label(speechesSample, logfile="/Users/blaette/Lab/tmp/foo.csv")
}
}

