% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pAttribute_method.R
\docType{methods}
\name{pAttribute,partition-method}
\alias{pAttribute,partition-method}
\title{pAttribute}
\usage{
\S4method{pAttribute}{partition}(object, from, to, mc = TRUE,
  verbose = TRUE)
}
\description{
pAttribute
}

