% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectLabel_function.R
\name{selectLabel}
\alias{selectLabel}
\title{Select Label}
\usage{
selectLabel(x, labelCol = "label")
}
\arguments{
\item{x}{Story to be told.}

\item{labelCol}{Column to use for labelling.}
}
\description{
Select label from data.frame!
}
