% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentenceSimilarity.R
\name{SentenceSimilarity}
\alias{SentenceSimilarity}
\title{Get similar sentences.}
\description{
Get similar sentences.

Get similar sentences.
}
\section{Methods}{

\describe{
\item{\code{getSentenceVectors(method = c("pos", "openNLP"), f = mean, stopwords = tm::stopwords("de"), tfidf = FALSE, verbose = TRUE)}}{}
\item{\code{}}{}
}
}

\examples{
\dontrun{
library(R6)
library(NLP)
library(openNLP)
library(openNLPmodels.de)
library(polmineR)

filename <- "/Users/blaette/Lab/tmp/word2vec/word2vec/data/plprbt_word2vec.csv"
W2V <- Word2Vec$new()
W2V$load(filename)
merkel <- partition("PLPRBT", speaker_name = "Angela Merkel", speaker_year = "2015")
S <- SentenceSimilarity$new()
S$partition <- merkel
S$word2vec <- W2V$vector_matrix
S$threads <- 3
S$getSentenceVectors(method = "pos", f = sum)
S$getSimilarityMatrix()

i <- 549
S$sentences[[i]]
sentence <- c("Kriminelle", "Ausländer", "raus", "und", "zwar", "schnell")
S$getSentenceSimilarity(sentence)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{partition}}{a partition}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SentenceSimilarity-getSentenceVectors}{\code{SentenceSimilarity$getSentenceVectors()}}
\item \href{#method-SentenceSimilarity-getSentenceVector}{\code{SentenceSimilarity$getSentenceVector()}}
\item \href{#method-SentenceSimilarity-getSentenceSimilarity}{\code{SentenceSimilarity$getSentenceSimilarity()}}
\item \href{#method-SentenceSimilarity-getSimilarityMatrix}{\code{SentenceSimilarity$getSimilarityMatrix()}}
\item \href{#method-SentenceSimilarity-getSimilarSentences}{\code{SentenceSimilarity$getSimilarSentences()}}
\item \href{#method-SentenceSimilarity-clone}{\code{SentenceSimilarity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SentenceSimilarity-getSentenceVectors"></a>}}
\if{latex}{\out{\hypertarget{method-SentenceSimilarity-getSentenceVectors}{}}}
\subsection{Method \code{getSentenceVectors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SentenceSimilarity$getSentenceVectors(
  method = c("pos", "openNLP"),
  f = mean,
  stopwords = tm::stopwords("de"),
  tfidf = FALSE,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SentenceSimilarity-getSentenceVector"></a>}}
\if{latex}{\out{\hypertarget{method-SentenceSimilarity-getSentenceVector}{}}}
\subsection{Method \code{getSentenceVector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SentenceSimilarity$getSentenceVector(sentence)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SentenceSimilarity-getSentenceSimilarity"></a>}}
\if{latex}{\out{\hypertarget{method-SentenceSimilarity-getSentenceSimilarity}{}}}
\subsection{Method \code{getSentenceSimilarity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SentenceSimilarity$getSentenceSimilarity(x, n = NULL, text = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SentenceSimilarity-getSimilarityMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-SentenceSimilarity-getSimilarityMatrix}{}}}
\subsection{Method \code{getSimilarityMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SentenceSimilarity$getSimilarityMatrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SentenceSimilarity-getSimilarSentences"></a>}}
\if{latex}{\out{\hypertarget{method-SentenceSimilarity-getSimilarSentences}{}}}
\subsection{Method \code{getSimilarSentences()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SentenceSimilarity$getSimilarSentences(i, n = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SentenceSimilarity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SentenceSimilarity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SentenceSimilarity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
