% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarity_method.R
\docType{methods}
\name{similarity,matrix_virtual_class-method}
\alias{similarity,matrix_virtual_class-method}
\title{importFrom proxy as.simil}
\usage{
\S4method{similarity}{matrix_virtual_class}(.Object, chunksize = 0,
  select = NULL, method = "cosine", progress = TRUE, verbose = TRUE,
  mc = FALSE)
}
\description{
importFrom proxy as.simil
}

