% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Word2Vec.R
\name{Word2Vec}
\alias{Word2Vec}
\title{Use word2vec.}
\description{
http://dsnotes.com/articles/glove-enwiki
}
\examples{
\dontrun{
filename <- "/Users/blaette/Lab/tmp/word2vec/word2vec/data/plprbt_word2vec.csv"
W2V <- Word2Vec$new()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Word2vec-load}{\code{Word2Vec$load()}}
\item \href{#method-Word2vec-similarity}{\code{Word2Vec$similarity()}}
\item \href{#method-Word2vec-clone}{\code{Word2Vec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Word2vec-load"></a>}}
\if{latex}{\out{\hypertarget{method-Word2vec-load}{}}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Word2Vec$load(filename, method = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Word2vec-similarity"></a>}}
\if{latex}{\out{\hypertarget{method-Word2vec-similarity}{}}}
\subsection{Method \code{similarity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Word2Vec$similarity(x, n = 50)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Word2vec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Word2vec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Word2Vec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
