% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.tibble_method.R
\name{as.tibble}
\alias{as.tibble}
\alias{as.tibble,partition-method}
\alias{as.tibble,partition_bundle-method}
\title{Turn partition into tibble.}
\usage{
as.tibble(.Object, ...)

\S4method{as.tibble}{partition}(.Object, p_attribute)

\S4method{as.tibble}{partition_bundle}(.Object, p_attribute, mc = getOption("polmineR.mc"))
}
\arguments{
\item{.Object}{An object the method is defined for.}

\item{...}{Further arguments.}

\item{p_attribute}{A positional attribute / p-attribute.}

\item{mc}{Number of cores to use.}
}
\description{
Turn partition into tibble.
}
\examples{
\dontrun{
library(polmineR)
library(tibble)
library(plyr)
library(dplyr)
library(pbapply)

plprbt <- partition("PLPRBT", speaker_type = "speech", speaker_year = "2010")
speeches <- as.speeches(
  plprbt,
  sAttributeDates = "speaker_date", sAttributeNames = "speaker_name", gap = 500,
  mc = TRUE, progress = TRUE
)
for (i in rev(which(summary(speeches)[["token"]] == 0))) speeches@objects[[i]] <- NULL
tib <- as.tibble(speeches, p_attribute = "word")
Y <- ddply(
  .data = tib,
  .variable = .(speaker_type, speaker_date, speaker_name, partition_name),
  .fun = function(x) paste(x[["word"]], collapse = " "),
  .progress = "text"
  )
colnames(Y)[5] <- "fulltext"
}
}
