% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize_method.R
\docType{methods}
\name{binarize}
\alias{binarize}
\alias{binarize,matrix-method}
\alias{binarize,simple_triplet_matrix-method}
\title{Binarize Matrix}
\usage{
binarize(.Object, ...)

\S4method{binarize}{simple_triplet_matrix}(.Object, top, verbose = TRUE)

\S4method{binarize}{matrix}(.Object, top, verbose = TRUE)
}
\arguments{
\item{.Object}{input object}

\item{top}{an integer value, the n values with the highes values that will be TRUE}

\item{verbose}{logical}
}
\description{
An input matrix (class matrix or simple_triplet_matrix) with numeric
values will be turned into a binary (logical) matrix. The top values
of each column will be assigned TRUE values.
}

