% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarity_method.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Calculate cosine Similarity}
\usage{
cosine_similarity(i, X, Y, ...)
}
\arguments{
\item{i}{integer index, the n-th position of the simple triplet matrix Y}

\item{X}{matrix from which the values are taken}

\item{Y}{simple triplet matrix}

\item{...}{further parameters (for use with blapply)}
}
\description{
Calculate cosine similarity of two vectors.
Note: For pairwise comparisons of vectors, this is much faster
than proxy::simil
}

