% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hitanalysis_class.R
\docType{data}
\name{Hitanalysis}
\alias{Hitanalysis}
\title{Analyse hits}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x6b6cf20> 
 - attr(*, "name")= chr "Hitanalysis_generator"
}}
\usage{
Hitanalysis
}
\description{
A story to be told.
}
\keyword{datasets}

