% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hitanalysis_class.R
\name{Hitanalysis}
\alias{Hitanalysis}
\title{Analyse hits}
\description{
A story to be told.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Hitanalysis-new}{\code{Hitanalysis$new()}}
\item \href{#method-Hitanalysis-read}{\code{Hitanalysis$read()}}
\item \href{#method-Hitanalysis-label}{\code{Hitanalysis$label()}}
\item \href{#method-Hitanalysis-clone}{\code{Hitanalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hitanalysis-new"></a>}}
\if{latex}{\out{\hypertarget{method-Hitanalysis-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hitanalysis$new(hits, bundle = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hitanalysis-read"></a>}}
\if{latex}{\out{\hypertarget{method-Hitanalysis-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hitanalysis$read(
  i,
  sAttribute = NULL,
  meta = NULL,
  highlight = NULL,
  type = "plpr",
  cqp = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hitanalysis-label"></a>}}
\if{latex}{\out{\hypertarget{method-Hitanalysis-label}{}}}
\subsection{Method \code{label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hitanalysis$label(
  categories = NULL,
  closed = TRUE,
  sAttribute = NULL,
  type = "plpr",
  highlight = NULL,
  continue = FALSE,
  cqp = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hitanalysis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Hitanalysis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hitanalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
