% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_method.R
\name{label}
\alias{label}
\alias{label,partition_bundle-method}
\title{assign labels}
\usage{
label(.Object, ...)

\S4method{label}{partition_bundle}(
  .Object,
  labels = c(true = 1, false = 0),
  description = "Make your choice",
  logfile = NULL,
  resume = FALSE,
  ...
)
}
\arguments{
\item{.Object}{an object of class \code{partition_bundle}}

\item{labels}{labels to assign}

\item{description}{output}

\item{logfile}{a logfile}

\item{resume}{logical}
}
\description{
assign labels
}
\examples{
\dontrun{
use("polmineR.sampleCorpus")
all <- partition("PLPRBTTXT", list(text_id=".*"), regex=TRUE, type="plpr")
speeches <- as.speeches(all, sAttributeDates="text_date", sAttributeNames="text_name", gap=500)
speechesSample <- sample(speeches, 25)
df <- label(speechesSample, logfile="/Users/blaette/Lab/tmp/foo.csv")
} 
}
