% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchars_method.R
\name{nchars}
\alias{nchars}
\alias{nchars,partition-method}
\alias{nchars,partition_bundle-method}
\title{Count the number of characters}
\usage{
nchars(x, ...)

\S4method{nchars}{partition}(
  x,
  p_attribute = "word",
  regexCharsToKeep = "[a-zA-Z]",
  toLower = TRUE,
  decreasing = TRUE,
  ...
)

\S4method{nchars}{partition_bundle}(x, decreasing = TRUE, mc = FALSE, progress = TRUE, ...)
}
\arguments{
\item{x}{Object to process.}

\item{...}{parameters that will be passed}

\item{p_attribute}{the p-attribute}

\item{regexCharsToKeep}{if NULL, counts for all charactrs will be returned, else a regex indicating which characters to include in the counting}

\item{toLower}{whether to lower tokens}

\item{decreasing}{logical, passed into order call}

\item{mc}{logical}

\item{progress}{A \code{logical} value.}
}
\description{
Count the number of characters

count characters
}
