% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectLabel_function.R
\name{selectLabel}
\alias{selectLabel}
\title{Select Label}
\usage{
selectLabel(x, labelCol = "label")
}
\arguments{
\item{categories}{a \code{data.frame} with cols category_no, and, optionally, subcategory_logical, and subcategory_no}

\item{descriptionCol}{character vector}
}
\description{
Select label from data.frame!
}
