% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_method.R
\name{similarity}
\alias{similarity}
\title{Similarity Calculations for Matrices}
\usage{
similarity(.Object, chunks = 1, select = NULL, method = "cosine",
  return.simil = TRUE, progress = TRUE, verbose = TRUE, mc = FALSE)
}
\arguments{
\item{.Object}{a simple_triplet_matrix}

\item{select}{a simple_triplet_matrix}

\item{method}{defaults to "cosine", no other method is implemented}

\item{progress}{logical}

\item{verbose}{logical}

\item{mc}{logical, or if numeric, providing number of cores}
}
\description{
Similarity Calculations for Matrices
}

