% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/virtualMatrixClass.R
\docType{methods}
\name{virtualMatrixClass}
\alias{blapply,virtualMatrixClass-method}
\alias{simple_triplet_matrix-class}
\alias{virtualMatrixClass}
\alias{virtualMatrixClass-class}
\title{virtualMatrixClass}
\usage{
\S4method{blapply}{virtualMatrixClass}(x, f, chunks = 2,
  finalizer = list(as.matrix, as.dist, as.matrix), progress = TRUE,
  verbose = FALSE, mc = FALSE, ...)
}
\arguments{
\item{chunks}{number of chunks to generate when splitting up an object for processing it in parallel}

\item{finalizer}{a list of functions that will be applied to the return value before the respective object is returned}
}
\description{
This is a virtual class to make it possible do define methods with uniform
behavior for the classes of the virtualMatrixClass.
}

